requirements.txt
attrs
branca
certifi
charset-normalizer
contourpy
cycler
folium
fonttools
geographiclib
geopy
h11
idna
Jinja2
kiwisolver
MarkupSafe
matplotlib
numpy
outcome
packaging
pillow
pyparsing
PyQt5
PyQt5-Qt5
PyQt5-sip
pyqtgraph
pyserial
PySocks
python-dateutil
requests
selenium
setuptools
six
sniffio
sortedcontainers
trio
trio-websocket
typing_extensions
urllib3
wheel
wsproto
xyzservices
